Ultima Versión: -Mod boton copiar al portapapeles - Añadida pestaña de confirmacion de borrado de descuento

+++ src/main/java/isw21/paginas/JInicio.java
@@ -13,11 +13,13 @@ import main.java.isw21.descuentos.Oferta;
 import main.java.isw21.descuentos.Porcentaje;
 import main.java.isw21.domain.Customer;
 import main.java.isw21.client.Client;
-
+import java.awt.datatransfer.StringSelection;
+import java.awt.Toolkit;
+import java.awt.datatransfer.Clipboard;
 
 public class JInicio extends JFrame
 {
-    Client cliente;
+	Client cliente;
 	ArrayList<Oferta> plOfertas;
 	private Timer timer;
 	public JFrame parent= new JFrame();
@@ -29,7 +31,7 @@ public class JInicio extends JFrame
 	{
 		//Establecemos la conexión y el dueño de los descuentos
 		this.customer = customer;
-        this.cliente = cliente;
+		this.cliente = cliente;
 
 		// DESCRIPCION DEL ENTORNO GRAFICO
 
@@ -38,7 +40,7 @@ public class JInicio extends JFrame
 		this.setLayout(new BorderLayout());
 		Font fuente1 = new Font("Serif", 1, 18);
 
-	//NORTE
+		//NORTE
 		JPanel pnlNorte = new JPanel();
 		pnlNorte.setBackground(new Color(255,255,255));
 		pnlNorte.setLayout(new GridLayout(1,3));
@@ -47,20 +49,20 @@ public class JInicio extends JFrame
 		l5.setIcon(new ImageIcon(imagen1.getImage().getScaledInstance(150, 60, Image.SCALE_SMOOTH)));
 		JLabel l6 = new JLabel("MIS DESCUENTOS");
 		l6.setFont(fuente1);
-      		JLabel l7 = new JLabel();
-	     	pnlNorte.setPreferredSize(new Dimension(600, 50));
+		JLabel l7 = new JLabel();
+		pnlNorte.setPreferredSize(new Dimension(600, 50));
 
-	      	pnlNorte.add(l5);
-	      	pnlNorte.add(l6);
-	      	pnlNorte.add(l7);
+		pnlNorte.add(l5);
+		pnlNorte.add(l6);
+		pnlNorte.add(l7);
 
 		// getDescuentos devuelve todos los descuentos del que el cliente es el dueño
-        plOfertas = getDescuentos(customer);
+		plOfertas = getDescuentos(customer);
 		//eliminamos el primero ya que siempre nos devuelve en la primera posicion un descuento generico.
-        plOfertas.remove(0);
+		plOfertas.remove(0);
 
 
-	//CENTRO
+		//CENTRO
 		//Establecemos el entorno gráfico dependiendo de los descuentos que tenga asociados un cliente
 		JPanel pnlCentro = new JPanel();
 		pnlCentro.setBackground(new Color(174,200,178));
@@ -71,62 +73,62 @@ public class JInicio extends JFrame
 		// Al iniciar la pestaña, se mostrarán los descuentos asociados a la cuenta
 
 		//Si no hay, se mostrará un mensaje: "En este momento no tienes descuentos".
-        if(plOfertas == null || l == 0){
+		if(plOfertas == null || l == 0){
 			JLabel lno = new JLabel("En este momento no tienes descuentos");
 			lno.setFont(fuente1);
 			pnlCentro.add(lno);
 			//pnlCentro.add(btnCrearDescuento);
-		    //Quitar el descuento inicial de bienvenida
-        }
+			//Quitar el descuento inicial de bienvenida
+		}
 		//En caso contrario visualizarán en el centro de la pestaña
 		else {
-            for (Oferta i : plOfertas) {
+			for (Oferta i : plOfertas) {
 				//Para cada descuento que tenga el usuario, se llamará a la funcion mastrar. La cual organiza los descuentos y los muestra al usuario
-                mostrarDescuento(i,pnlCentro,fuente1);
+				mostrarDescuento(i,pnlCentro,fuente1);
 
 
-            }
-        }
+			}
+		}
 
 
 
-	//SUR
+		//SUR
 
-	    JPanel pnlSur = new JPanel();
-	    pnlSur.setLayout(new GridLayout(1, 5));
+		JPanel pnlSur = new JPanel();
+		pnlSur.setLayout(new GridLayout(1, 5));
 
-	    JButton btnMas = new JButton();
-	    btnMas.setBorder(BorderFactory.createLineBorder(Color.BLACK));
-	    ImageIcon imagenMas = new ImageIcon("src/main/java/isw21/media/Mas.png");
-	    btnMas.setIcon(new ImageIcon(imagenMas.getImage().getScaledInstance(50, 50, Image.SCALE_SMOOTH)));
+		JButton btnMas = new JButton();
+		btnMas.setBorder(BorderFactory.createLineBorder(Color.BLACK));
+		ImageIcon imagenMas = new ImageIcon("src/main/java/isw21/media/Mas.png");
+		btnMas.setIcon(new ImageIcon(imagenMas.getImage().getScaledInstance(50, 50, Image.SCALE_SMOOTH)));
 
-	    JButton btnLupa = new JButton();
-	    btnLupa.setBorder(BorderFactory.createLineBorder(Color.BLACK));
-	    ImageIcon imagenLupa = new ImageIcon("src/main/java/isw21/media/Lupa.png");
-	    btnLupa.setIcon(new ImageIcon(imagenLupa.getImage().getScaledInstance(50, 50, Image.SCALE_SMOOTH)));
+		JButton btnLupa = new JButton();
+		btnLupa.setBorder(BorderFactory.createLineBorder(Color.BLACK));
+		ImageIcon imagenLupa = new ImageIcon("src/main/java/isw21/media/Lupa.png");
+		btnLupa.setIcon(new ImageIcon(imagenLupa.getImage().getScaledInstance(50, 50, Image.SCALE_SMOOTH)));
 
-	    JButton btnHome = new JButton();
-	    btnHome.setBorder(BorderFactory.createLineBorder(Color.BLACK));
-	    ImageIcon imagenHome = new ImageIcon("src/main/java/isw21/media/Home.png");
-	    btnHome.setIcon(new ImageIcon(imagenHome.getImage().getScaledInstance(50, 50, Image.SCALE_SMOOTH)));
+		JButton btnHome = new JButton();
+		btnHome.setBorder(BorderFactory.createLineBorder(Color.BLACK));
+		ImageIcon imagenHome = new ImageIcon("src/main/java/isw21/media/Home.png");
+		btnHome.setIcon(new ImageIcon(imagenHome.getImage().getScaledInstance(50, 50, Image.SCALE_SMOOTH)));
 
-	    JButton btnCorazon = new JButton();
-	    btnCorazon.setBorder(BorderFactory.createLineBorder(Color.BLACK));
-	    ImageIcon imagenCorazon = new ImageIcon("src/main/java/isw21/media/Corazon.png");
-	    btnCorazon.setIcon(new ImageIcon(imagenCorazon.getImage().getScaledInstance(50, 50, Image.SCALE_SMOOTH)));
+		JButton btnCorazon = new JButton();
+		btnCorazon.setBorder(BorderFactory.createLineBorder(Color.BLACK));
+		ImageIcon imagenCorazon = new ImageIcon("src/main/java/isw21/media/Corazon.png");
+		btnCorazon.setIcon(new ImageIcon(imagenCorazon.getImage().getScaledInstance(50, 50, Image.SCALE_SMOOTH)));
 
-	    JButton btnMiPerfil= new JButton();
-	    btnMiPerfil.setBorder(BorderFactory.createLineBorder(Color.BLACK));
-	    ImageIcon imagen = new ImageIcon("src/main/java/isw21/media/Perfil.png");
-	    btnMiPerfil.setIcon(new ImageIcon(imagen.getImage().getScaledInstance(70, 70, Image.SCALE_SMOOTH)));
+		JButton btnMiPerfil= new JButton();
+		btnMiPerfil.setBorder(BorderFactory.createLineBorder(Color.BLACK));
+		ImageIcon imagen = new ImageIcon("src/main/java/isw21/media/Perfil.png");
+		btnMiPerfil.setIcon(new ImageIcon(imagen.getImage().getScaledInstance(70, 70, Image.SCALE_SMOOTH)));
 
-	    pnlSur.add(btnMas);
-	    pnlSur.add(btnLupa);
-	    pnlSur.add(btnHome);
-	    pnlSur.add(btnCorazon);
-	    pnlSur.add(btnMiPerfil);
+		pnlSur.add(btnMas);
+		pnlSur.add(btnLupa);
+		pnlSur.add(btnHome);
+		pnlSur.add(btnCorazon);
+		pnlSur.add(btnMiPerfil);
 
-	//LADOS
+		//LADOS
 		JPanel pnlEste = new JPanel();
 		pnlEste.setBackground(new Color(174,200,178));
 		pnlEste.setPreferredSize(new Dimension(60, 60));
@@ -174,44 +176,57 @@ public class JInicio extends JFrame
 
 	// Metodo de visualización de descuentos
 	// Se ejecutará para los descuentos que tenga el usuario
-    private void mostrarDescuento(Oferta i , JPanel pnlCentro, Font fuente1) {
+	private void mostrarDescuento(Oferta i , JPanel pnlCentro, Font fuente1) {
 		String tipo = comprobarTipo(i);
-        JPanel pnlDescuento = new JPanel();
+		JPanel pnlDescuento = new JPanel();
 		pnlDescuento.setLayout(new GridLayout(6, 1));
-        if (i instanceof ChequeRegalo) {
+		if (i instanceof ChequeRegalo) {
 			pnlDescuento.setLayout(new GridLayout(7, 1));
 			ChequeRegalo cheque = (ChequeRegalo) i;
 			JLabel lblGastado = new JLabel("Gastado: " + cheque.getGastado(), JLabel.CENTER);
 		}
-        JLabel lblComercio = new JLabel(i.getComercio(), JLabel.CENTER);
-        JLabel lblValor = new JLabel("Valor: " + i.getValor(), JLabel.CENTER);
-        JLabel lblTipo = new JLabel("Tipo: " + tipo, JLabel.CENTER);
-        JLabel lblCodigo = new JLabel("Codigo: " + i.getCodigo(), JLabel.CENTER);
-        JLabel lblFechaFin = new JLabel("Caduca: " + i.getFechaFin(),JLabel.CENTER);
-		JButton btnInfo = new JButton("Info del decuento");
+		JLabel lblComercio = new JLabel(i.getComercio(), JLabel.CENTER);
+		JLabel lblValor = new JLabel("Valor: " + i.getValor(), JLabel.CENTER);
+		JLabel lblTipo = new JLabel("Tipo: " + tipo, JLabel.CENTER);
+		JLabel lblCodigo = new JLabel("Codigo: " + i.getCodigo(), JLabel.CENTER);
+		JLabel lblFechaFin = new JLabel("Caduca: " + i.getFechaFin(),JLabel.CENTER);
+		JButton btnInfo = new JButton("Info del descuento");
+		JButton btnCopiar = new JButton("Copiar");
 		JButton btnUsar = new JButton("Usar descuento");
-        JButton btnEliminar = new JButton("Eliminar descuento");
+		JButton btnEliminar = new JButton("Eliminar descuento");
 
-        lblComercio.setBorder(BorderFactory.createEtchedBorder());
-        //new JLabel( "prueba",JLabel.CENTER
+		lblComercio.setBorder(BorderFactory.createEtchedBorder());
+		//new JLabel( "prueba",JLabel.CENTER
 		JPanel pnlBotones = new JPanel(new GridLayout(1,3));
 
 
-        pnlDescuento.add(lblComercio);
-        pnlDescuento.add(lblValor);
+		pnlDescuento.add(lblComercio);
+		pnlDescuento.add(lblValor);
 		if (i instanceof ChequeRegalo) {
 			pnlDescuento.setLayout(new GridLayout(7, 1));
 			ChequeRegalo cheque = (ChequeRegalo) i;
 			JLabel lblGastado = new JLabel("Gastado: " + cheque.getGastado(), JLabel.CENTER);
 			pnlDescuento.add(lblGastado);
 		}
-        pnlDescuento.add(lblTipo);
-        pnlDescuento.add(lblCodigo);
-        pnlDescuento.add(lblFechaFin);
+		pnlDescuento.add(lblTipo);
+		pnlDescuento.add(lblCodigo);
+		pnlDescuento.add(lblFechaFin);
 		pnlBotones.add(btnInfo);
+		pnlBotones.add(btnCopiar);
 		pnlBotones.add(btnUsar);
 		pnlBotones.add(btnEliminar);
-        pnlDescuento.add(pnlBotones);
+		pnlDescuento.add(pnlBotones);
+
+		btnCopiar.addActionListener(new ActionListener() {
+			@Override
+			public void actionPerformed(ActionEvent e) {
+				StringSelection stringSelection = new StringSelection(i.getCodigo());
+				Clipboard clipboard = Toolkit.getDefaultToolkit().getSystemClipboard();
+				clipboard.setContents(stringSelection, null);
+				JOptionPane.showMessageDialog(null,"Se ha copiado el código del descuento al portapapeles.");
+			}
+		});
+
 
 		btnInfo.addActionListener(new ActionListener() {
 			@Override
@@ -225,57 +240,63 @@ public class JInicio extends JFrame
 		btnEliminar.addActionListener(new ActionListener()
 		{
 			public void actionPerformed(ActionEvent e)
+
 			{
-				// Si seleccionamos la opcion de crer descuento, abrimos el entorno gráfico necesario y
-				// le introducimos como parámetros, el dueño, la conexion con el servidor y los descuentos asociados al dueño.
-				String eliminado= i.getComercio();
-				OfertaDAO.eliminarDescuento(customer,i);
-				System.out.println("Descuento eliminado");
-				pnlCentro.removeAll();
-				pnlCentro.setVisible(false);
-				plOfertas = getDescuentos(customer);
-				plOfertas.remove(0);
-				int l= plOfertas.size();
-				//Esto habrá que cambiarlo, pero de momento con 16 está nice
-				pnlCentro.setLayout(new GridLayout(4, 4));
+				int dialogResult = JOptionPane.showConfirmDialog(null,"¿Desea borrar el descuento de "+i.getComercio()+"?");
 
-				// Al iniciar la pestaña, se mostrarán los descuentos asociados a la cuenta
+				if (dialogResult == JOptionPane.YES_OPTION){
+					String eliminado= i.getComercio();
+					OfertaDAO.eliminarDescuento(customer,i);
+					System.out.println("Descuento eliminado");
+					pnlCentro.removeAll();
+					pnlCentro.setVisible(false);
+					plOfertas = getDescuentos(customer);
+					plOfertas.remove(0);
+					int l= plOfertas.size();
+					//Esto habrá que cambiarlo, pero de momento con 16 está nice
+					pnlCentro.setLayout(new GridLayout(4, 4));
 
-				//Si no hay, se mostrará un mensaje: "En este momento no tienes descuentos".
-				if(plOfertas == null || l == 0	){
+					// Al iniciar la pestaña, se mostrarán los descuentos asociados a la cuenta
+
+					//Si no hay, se mostrará un mensaje: "En este momento no tienes descuentos".
+					if(plOfertas == null || l == 0	){
 						JLabel lno = new JLabel("En este momento no tienes descuentos");
 						lno.setFont(fuente1);
 						pnlCentro.add(lno);
 						//pnlCentro.add(btnCrearDescuento);
 						//Quitar el descuento inicial de bienvenida
 					}
-				//En caso contrario visualizarán en el centro de la pestaña
-				else {
-					for (Oferta i : plOfertas) {
-						//Para cada descuento que tenga el usuario, se llamará a la funcion mastrar. La cual organiza los descuentos y los muestra al usuario
-						mostrarDescuento(i,pnlCentro,fuente1);
+					//En caso contrario visualizarán en el centro de la pestaña
+					else {
+						for (Oferta i : plOfertas) {
+							//Para cada descuento que tenga el usuario, se llamará a la funcion mastrar. La cual organiza los descuentos y los muestra al usuario
+							mostrarDescuento(i,pnlCentro,fuente1);
 
 
+						}
 					}
+					pnlCentro.setVisible(true);
+					JOptionPane.showMessageDialog(parent,"Se ha eiliminado el descuento de "+eliminado+" y todos los descuentos duplicados.");
+
 				}
-				pnlCentro.setVisible(true);
-				JOptionPane.showMessageDialog(parent,"Se ha eiliminado el descuento de "+eliminado+" y todos los descuentos duplicados.");
+				// Si seleccionamos la opcion de crer descuento, abrimos el entorno gráfico necesario y
+				// le introducimos como parámetros, el dueño, la conexion con el servidor y los descuentos asociados al dueño.
 
 				//mostrarDescuento(plDescuentos.get(plDescuentos.size()-1),pnlCentro);
 
 			}
 		});
 
-        pnlDescuento.setAlignmentX(lblComercio.CENTER_ALIGNMENT);
-        pnlDescuento.setAlignmentX(lblValor.CENTER_ALIGNMENT);
-        pnlDescuento.setAlignmentX(lblTipo.CENTER_ALIGNMENT);
-        pnlDescuento.setAlignmentX(lblCodigo.CENTER_ALIGNMENT);
+		pnlDescuento.setAlignmentX(lblComercio.CENTER_ALIGNMENT);
+		pnlDescuento.setAlignmentX(lblValor.CENTER_ALIGNMENT);
+		pnlDescuento.setAlignmentX(lblTipo.CENTER_ALIGNMENT);
+		pnlDescuento.setAlignmentX(lblCodigo.CENTER_ALIGNMENT);
 		pnlDescuento.setAlignmentX(lblFechaFin.CENTER_ALIGNMENT);
 
-        pnlDescuento.setBorder(BorderFactory.createEtchedBorder());
+		pnlDescuento.setBorder(BorderFactory.createEtchedBorder());
 
-        pnlCentro.add(pnlDescuento);
-    }
+		pnlCentro.add(pnlDescuento);
+	}
 
 	private String comprobarTipo(Oferta i) {
 		if (i instanceof Descuento){
@@ -290,24 +311,24 @@ public class JInicio extends JFrame
 	}
 
 	// Metodo que obtiene los descuentos de cada cliente.
-    public ArrayList<Oferta> getDescuentos(Customer customer)
+	public ArrayList<Oferta> getDescuentos(Customer customer)
 	{
 		//Hacemos la solicitud a la base de datos medainte un mensaje con contexto de getDescuentos.
-        cliente.setContext("/getDescuentos");
-        HashMap<String,Object> session= new HashMap<String,Object>();
+		cliente.setContext("/getDescuentos");
+		HashMap<String,Object> session= new HashMap<String,Object>();
 		// Los campos que tendrá el mensaje de salida serán el customer --> dueño del descuento
 		// Y un arraylist donde queremos que se guarden los descuentos. Para el correcto funcionamiento, pondremos
 		// dentro del arraylist un descuento generico para no tener problemas con la lista vacia
-        session.put("Customer",customer);
-        ArrayList<Oferta> ofertas = new ArrayList<Oferta>();
-        ofertas.add(new Descuento("Diskant (Bienvenida)","Hoy","Nunca",50,"DISKANTMOLA"));
-        session.put("Descuentos", ofertas);
-        cliente.setSession(session);
+		session.put("Customer",customer);
+		ArrayList<Oferta> ofertas = new ArrayList<Oferta>();
+		ofertas.add(new Descuento("Diskant (Bienvenida)","Hoy","Nunca",50,"DISKANTMOLA"));
+		session.put("Descuentos", ofertas);
+		cliente.setSession(session);
 		//Mandamos la comunicacion al servidor
-        cliente.run(cliente);
+		cliente.run(cliente);
 		// Una vez hecha la conexion, el cliente ya tiene los descuentos del usuario
-        return cliente.getDescuentos();
-    }
+		return cliente.getDescuentos();
+	}
 
 
 }
